Added support for simple urls for static content.
Files placed under the static directory (by default, is configurable
as an array of possible directories under the toplevel STATIC) are
accessible from the root directory.
For example if you place a test.html in the static directory, in your
browser go to SITE_URL/test.html, if you don't have an explicit route
that handles test.html the system will look in the static/ dir for
any files named test.html. In this case it finds it so includes it.
If the file is not found the 404 handler is called as normal. Only
html, htm and php files are executed, for other files a redirect to the
actual location is returned by the router. This is for security as
it prevents direct access to executable php code. To get around the
direct php call check you simple leave the php extension off the url.
So you have the file static/contact.php, to access you would go to
SITE_URL/contact. The php file is included into the router so you
have access to all the framework and can use templates and whatnot.