Extended Hooks system so it can work like wordpress'
Changed plugin loading behaviour. In the modules folder
it will continue to load files in the root. In sub-folders it
will first look for a file named the same as the folder + '.php' and load
it only or it will load all files in the directory like
the old behaviour.